fewer warnings about precision loss from type conversion
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@3133 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/limn/light.c
@@ -70,7 +70,7 @@ limnLightAmbientSet(limnLight *lit, float r, float g, float b) {
 int
 limnLightUpdate(limnLight *lit, limnCamera *cam) {
   char me[]="limnLightUpdate", err[BIFF_STRLEN];
-  float dir[3], _dir[3], uvn[9]={0,0,0,0,0,0,0,0,0}, norm;
+  double dir[3], _dir[3], uvn[9]={0,0,0,0,0,0,0,0,0}, norm;
   int i;
   
   if (cam) {
